Basic IK movement (#303)
* RE:SS3D Icon on the .exe

Added the RE:SS3D icon to the project, it got lost somehow.
Changed the slot UI to something more brighter, altho it'll need a complete rework since it doesn't support proper scalling.

* More smoothing to the movement

Added a little cubic function to smooth the input.
On the InventoryUI prefab there's a small change from the legacy text component to the TextMeshPro component

* Basic Humanoid Chest IK

Changes on the character movement script.
Renamed to HumanoidMovementController, added ChestIK and ChestBone.
Added empties for the human IK on it's prefab

* Revert "Basic Humanoid Chest IK"

This reverts commit bef26f67111f9633ffdb4b14d00a980cb63209ac.

* Revert "Revert "Basic Humanoid Chest IK""

This reverts commit 60ec5cc67895cfb862e589d85aaee8346b6e885c.

* Revert "Revert "Revert "Basic Humanoid Chest IK"""

This reverts commit 10ca3fa946de922c8fbc263b94f0fdefefbd7469.

* Basic IK and Rotation limits

This adds chest and head IK with limits on their rotation.
(this code needs cleaning)

* Update manifest.json

* Cleaning HumanoidMovementController.cs

Renamed MovementController to HumanoidMovementController.
This cleans a little the code from the HumanoidMovementController, which was pretty much a bunch of code noodle.
Human prefab changes is just the script.

* Update HumanoidMovementController.cs

Changed Update to FixedUpdate.
Changed the magnitude checks to Vector2.zero checks.
Changed LerpUnclamped to Lerp.

* Movement and UI fixes

Fixed all the floor prefabs to have convex enabled on the mesh collider.
All the UI changes were Pixel Per Unit on 3 to add more quality to the panels.
The HumanoidMovementController now has a little better IK and more control over the rotations, aswell as fixing the bug with walking, stopping, and pressing Q or E on the same frame causing the character to move towards that position, also fixed the head rotation caused by that same thing.
Changed the DelayedInteraction's loading bar spawn to up because it spawns on the ground.

* Fixed some floor instances missing convex
