+++ Assets/Content/Structures/Floors/Reinforced Floor/ReinforcedFloor.prefab
@@ -26,7 +26,7 @@ PrefabInstance:
     - target: {fileID: 5870127389560564871, guid: 8686fbdb413e33b4395a775c4f2ba990,
         type: 3}
       propertyPath: m_Convex
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 6745463052541861222, guid: 8686fbdb413e33b4395a775c4f2ba990,
         type: 3}
